requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
cohere==5.0.0
qdrant-client
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
trafilatura
tqdm==4.66.1
